requirements.txt
discord.py>=2.6.0
aiohttp
python-dotenv